refactor(message): Message 클래스 package 변경
