+++ src/main/java/lotto/global/exception/CustomException.java
@@ -1,5 +1,7 @@
 package lotto.global.exception;
 
+import lotto.global.message.ErrorMessage;
+
 public class CustomException extends IllegalArgumentException {
     private CustomException(ErrorMessage errorMessage) {
         super(errorMessage.getMessage());
