+++ src/main/java/lotto/global/message/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto.global.exception;
+package lotto.global.message;
 
 public enum ErrorMessage {
     BLANK_INPUT("빈 문자열이 입력되었습니다."),
